gemini_cli_mcp/pyproject.toml
[project]
name = "gemini-cli-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "mcp>=1.10.1",
    "google-genai>=1.23.0",
    "fastmcp>=2.12.4",
]
